Add function createGrid() with hover effect
This function will create a certain number
of div depending on the function arguments.
Thsi function calculates the width and heigth
for the divs. this function also adds an
eventlistener to each div so that when is hovered,
a class called "test" will be added to the div,
which will change the background color.